FrictionSim2D_project/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "2DFrictionSim"
version = "0.1.0"
description = "A tool for generating LAMMPS input scripts for 2D material friction simulations."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Matteo Valderrama", email = "m.valderrama23@imperial.ac.uk"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
  "numpy>=1.21",
  "ase>=3.22",
  "lammps>=2023.8.2",  # Python bindings for LAMMPS
  "PyYAML>=6.0",
  "jinja2>=3.1",
  "pydantic>=2.0",
]

[project.scripts]
2DFrictionSim = "2DFrictionSim.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/2DFrictionSim"]